requirements.txt
google-cloud-bigquery==3.38.0
google-auth==2.42.0
google-cloud-core==2.4.3
pandas==2.3.3
numpy==2.3.4
pyarrow==22.0.0
requests==2.32.5
python-dotenv==1.2.1
python-dateutil==2.9.0.post0

